% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestools.R
\name{bt_check_convergence}
\alias{bt_check_convergence}
\title{Check convergence of BayesianTools output}
\usage{
bt_check_convergence(samples, threshold = 1.1, use_CI = TRUE,
  use_mpsrf = TRUE)
}
\description{
Check convergence of BayesianTools output
}
