% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process-output.R
\name{burnin.thin}
\alias{burnin.thin}
\title{Burn-in and thinning of MCMC samples}
\usage{
burnin.thin(samples, target = 5000, burnin.ratio = 2, auto = TRUE,
  burnin = NULL, thin = NULL)
}
\arguments{
\item{samples}{Matrix of MCMC samples}

\item{target}{Target number of samples (default = 5000). Only applicable if
auto=TRUE.}

\item{burnin.ratio}{Fraction of samples to burn-in; i.e. 2 means to remove
first 1/2 of samples, 3 means 1/3, etc. (default = 2). Only applicable if
auto=TRUE.}

\item{auto}{Whether or not to perform automatic burnin and thin based on
target number of samples.}

\item{burnin}{Number of samples to discard as burnin (auto must be FALSE)}

\item{thin}{Thinning interval (auto must be FALSE)}
}
\description{
Burn-in and thinning of MCMC samples
}

