% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvt.R
\name{bvt}
\alias{bvt}
\alias{rbvt}
\alias{dbvt}
\title{Bivariate \eqn{t}-distributions}
\usage{
rbvt(n, df, theta)

dbvt(t1, t2, df, theta, cutoff, truncation = NULL)
}
\arguments{
\item{n}{the number of samples required}

\item{df}{a numeric vector giving the degree of freedoms of variables}

\item{theta}{a numeric value of mixing angle}

\item{t1}{a numeric vector of baseline measurements}

\item{t2}{a numeric vector of followup measurements}

\item{cutoff}{a numeric value of cutoff point in which the baseline measurements are above or below}

\item{truncation}{a character string; one of \code{"left"} or \code{"right"} indicating the specification of distribution.}
}
\value{
It give the matrix of random numbers and density whose dimension is determine by \code{n}, \eqn{t_1}, and \eqn{t_2}.
}
\description{
Generating random numbers, and density from bivariate \eqn{t}-distributions either truncated or complete with \code{df} degree of freedom neither necessary equal
nor integer and mixing angle \eqn{\theta}.
}
\details{
The bivariate \eqn{t}-distributions have identical or unequal marginals and equal or unequal degree of freedoms proposed by Shaw, W. T., & Lee, K. T. A. (2008).
The density function of bivariate \eqn{t}-distribution with equal degree of freedom is
\deqn{f\left( {t_1},{t_2};n,\theta \right) = \frac{1}{{2\pi \cos \left( \theta  \right)}}\frac{1}{{{{\left( {1 + \frac{\Delta }{n}} \right)}^{n/2 + 1}}}}}
where \eqn{-\infty < {t_1} < \infty}, \eqn{- \infty  < {t_2} < \infty}, \eqn{\Delta  = ({t_1^2 + t_2^2 - 2{t_1}{t_2}\sin (\theta )}){{{\cos }^{-2}}(\theta )}}, \eqn{n} is the degree of freedom, \eqn{\theta} is the mixing angle.
The test statistics is equal to \eqn{{T_1} = \sqrt {n/{C^2}} {Z_1}} and \eqn{{T_2} = \sqrt {n/{C^2}} \left( {{Z_1}\sin \left( \theta  \right) + {Z_2}\cos \left( \theta  \right)} \right)} where \eqn{\left( {{Z_1},{Z_2}} \right)\sim N\left( {0,1} \right)} and \eqn{{C^2} \sim \chi _n^2}.
The mean of \eqn{t_1} and \eqn{t_2} is equal to 0 and variance \eqn{\frac{n}{n-2}}. If the degree of freedom for \eqn{t_1} and \eqn{t_2} is different with independent denominator then the test statistics are defined as
\eqn{{T_1} = \sqrt {{n_1}/z} {Z_1}}, \eqn{{T_2} = \left( {{Z_2}\cos \left( \theta  \right) + \sin \left( \theta  \right)\sqrt {z/{n_1}} {T_1}} \right)/\sqrt {w/{n_2}}}.
Where \eqn{n_1}, \eqn{n_2} are the degree of freedoms, \eqn{z \sim \chi _{{n_1}}^2}, \eqn{w \sim \chi _{{n_2}}^2} and \eqn{\left( {{Z_1},{Z_2}} \right) \sim N\left( {0,1} \right)}. The density is derived by authors through loosely grouped approach.
}
\examples{
  ### Generating a 1,000 samples
randomT<-rbvt(n=1000,df=c(15,9),theta=pi/4)
hist(randomT[,1])

  #create bivariate t-distribution
t1    <- sort(randomT[,1])
t2    <- sort(randomT[,2])
df    <- c(15,9)
f     <- function(t1, t2) dbvt(t1,t2,df=c(15,9),theta= pi/4)
z     <- outer(t1, t2, f)

  #create contour plot
contour(t1, t2, z)

  ### Compute density from truncated distribution
randTrunc<-randomT[randomT[,1]>=0.5,]
  ### create right trucatated bivariate t-distribution
t1.Tr <- sort(randTrunc[,1])
t2.Tr <- sort(randTrunc[,2])
df    <- c(15,9)
f     <- function(t1.Tr, t2.Tr) dbvt(t1.Tr,t2.Tr,
df=c(15,9),theta= pi/4,cutoff = 0.5,truncation = "right")
z     <- outer(t1.Tr, t2.Tr, f)

  #create contour plot
contour(t1.Tr, t2.Tr, z)
}
\references{
Shaw, W. T., & Lee, K. T. A. (2008). Bivariate Student t distributions
with variable marginal degrees of freedom and independence. \emph{Journal of Multivariate Analysis}, 99(6), 1276-1287.
}
\seealso{
\code{\link{RTM.T}}
}
\author{
Shaw, W. T., & Lee, K. T. A.
}
