% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.spectra.R
\name{cbind.spectra}
\alias{cbind.spectra}
\title{Combine spectra by wavelength}
\usage{
\method{cbind}{spectra}(...)
}
\arguments{
\item{...}{Spectra to combine}
}
\description{
Combine spectra by wavelength
}
