% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.convergence.R
\name{check.convergence}
\alias{check.convergence}
\title{Check convergence of multiple MCMC chains}
\usage{
check.convergence(jags_out, threshold = 1.1, verbose = TRUE, ...)
}
\arguments{
\item{jags_out}{mcmc.list object (from coda package) containing
samples from MCMC chains.}

\item{threshold}{Gelman-Rubin diagnostic parameter threshold. Default = 1.1}

\item{verbose}{If TRUE, print convergence result. Default = TRUE}

\item{...}{Additional arguments to \code{gelman.diag} (from coda package)}
}
\value{
List length 3 containing the following:
* convergence: Logical. Whether or not convergence was achieved.
* diagnostics: Numerical value of Gelman-Rubin diagnostics for each parameter and multivariate diagnostic
* error: Logical. Whether or not an error occured in the Gelman-Rubin calculation.
}
\description{
Uses Gelman multivariate Gelman-Rubin diagnostic to check if
multiple MCMC chains have converged
}
