% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neff.R
\name{corr_max_lag}
\alias{corr_max_lag}
\title{Calculate max ACF lag from correlation power analysis}
\usage{
corr_max_lag(nx, r = 0.1, sig.level = 0.05, power = 0.95, ...)
}
\description{
Calculate max ACF lag from correlation power analysis
}
