% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataspec.R
\name{dataspec_4}
\alias{dataspec_4}
\alias{dataspec_5}
\alias{dataspec_d}
\alias{refractive_45}
\alias{refractive_d}
\title{PROSPECT model driver data}
\usage{
dataspec_4()

dataspec_5()

dataspec_d()

refractive_45()

refractive_d()
}
\value{
For \code{refractive}, a 2101 (400-2500 nm) vector of refractive
index values. For \code{dataspec} a matrix (2101 x N) of specific
absorption coefficients.
}
\description{
\code{dataspec_X} returns the specific absorption coefficients for
PROSPECT parameters of the corresponding version. \code{refractive_X}
returns the refractive index values.
}
