% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/default.invert.prospect.R
\name{default.invert.prospect}
\alias{default.invert.prospect}
\title{Default settings for PROSPECT inversion}
\usage{
default.invert.prospect(observed, sensor, ngibbs, version = 5,
  do.mle = TRUE, quiet = TRUE)
}
\arguments{
\item{observed}{Vector of observed reflectance.}

\item{sensor}{Sensor name (string)}

\item{ngibbs}{Number of MCMC iterations (integer)}

\item{version}{PROSPECT version (4, 5 [default], or 5B}

\item{do.mle}{Whether or not to perform LM optimization of initial
conditions before starting Metropolis-Hastings sampling (default = TRUE)}

\item{quiet}{If TRUE, don't print status updates or LM fit results (default
= TRUE)}
}
\description{
Bayesian inversion of PROSPECT with default priors and
parameter constraints and random initial conditions.
}

