% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\docType{data}
\name{default.settings.prospect}
\alias{default.settings.prospect}
\title{Defult inversion settings for PROSPECT 5 models}
\format{
An object of class \code{list} of length 18.
}
\usage{
default.settings.prospect
}
\description{
Default settings for PROSPECT inversion
}
\keyword{datasets}
