% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truncnorm.R
\name{dtnorm}
\alias{dtnorm}
\title{Truncated normal distribution density}
\usage{
dtnorm(x, mu, sd, MIN)
}
\arguments{
\item{x}{A random variable}

\item{mu}{The mean parameter of the distribution; NOTE this is not equal to the mean}

\item{sd}{The standard deviation parameter of the distribution}

\item{MIN}{Value at which the truncation takes place}
}
\value{
Numeric; log density of the distribution, or -1e15 if the x < MIN
}
\description{
Truncated normal distribution density
}
\details{
Calculates the log density of a univariate truncated normal variable
}
\author{
Alexey Shiklomanov
}

