% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.wrapper.R
\name{EDR.preprocess.ed2in}
\alias{EDR.preprocess.ed2in}
\title{Preprocess ED2IN file for EDR}
\usage{
EDR.preprocess.ed2in(ed2in.path, output.path, datetime, history.full.prefix)
}
\arguments{
\item{ed2in.path}{Path to original ED2IN file.}

\item{output.path}{Path to directory for new ED2IN file (and where analysis 
is performed)}

\item{datetime}{POSIX datetime object defining the time at which to run EDR}

\item{history.full.prefix}{Full path and prefix for history file}
}
\description{
Copy ED2IN to `output.path` and modify for EDR specifications
}

