% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.wrapper.R
\name{EDR}
\alias{EDR}
\title{ED radiative transfer module (EDR) wrapper function

This function provides a convenient way to call the ED
radiative transfer module (EDR, which simulates full spectral return of an
ED patch for a given point in time) directly from R.}
\usage{
EDR(paths, spectra_list, par.wl, nir.wl, datetime, trait.values,
  settings = list(model = list(revision = "git", config.header = NULL)),
  history.prefix = "history", edr.exe.name = "ed_2.1-opt",
  change.history.time = TRUE, output.path = getwd(), clean = FALSE)
}
\arguments{
\item{paths}{List of relevant paths. Must contain the following:
\itemize{
\item \code{ed2in} -- Path to ED2IN file for the original ED run. If this is \code{NA}, the
function assumes that the ED2IN file in \code{output.path} is already up-to-date
and skips the analysis step.
\item \code{history} -- Path and prefix for history file for the run of interest;
\item \code{edr.exe} -- Path to EDR executable
\item \code{soil_reflect} -- Path to soil reflectance. If NULL or unset, use
reflectance in package. If NA, skip (assume data already in directory).
\item \code{wood_reflect} -- Path to wood reflectance. If NULL or unset, use
reflectance in package. If NA, skip (assume data already in directory).
}}

\item{spectra_list}{List of spectral data matrices. Names must exactly match
the PFTs given in \code{trait.values}. Each item must be a matrix of
reflectance (labeled 'R' or 'reflectance') and transmittance (labeled 'T' or
'transmittance') values. If the matrix is not of class \code{spectra} (see
\code{\link[=spectra]{spectra()}}), it must also have a column of wavelength values (labeled
'wl'). Such a matrix is returned by default by all versions of PROSPECT in
this package.}

\item{par.wl}{Vector of wavelengths defining PAR region}

\item{nir.wl}{Vector of wavelengths defining NIR region}

\item{datetime}{POSIXlt object defining the date and at which the run will
take place. N datetime POSIXlt object defining the date and at which the run
will take place. Note that runs at night and during the winter can give poor
results.}

\item{trait.values}{Named, hierarchical list of trait values for generating
config.xml file. Names must be PFTs, and must exactly match names of
\code{spectra_list}.}

\item{settings}{PEcAn settings list. Default is model$revision = 'git', model$config.header = NULL.}

\item{history.prefix}{Prefix in histroy file name. Will be appended to
history path.}

\item{edr.exe.name}{Name of EDR executable. Default = 'ed_2.1-opt'}

\item{change.history.time}{Logical. If \code{TRUE}, rename the history file to
the time specified in \code{datetime}. Default = TRUE.}

\item{output.path}{Path to store all output files. Default is current
working directory.}

\item{clean}{Logical. If \code{TRUE}, remove all files generated by this function
(e.g. cloned history file, ED2IN, output HDF files).}
}
