% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sail.R
\name{foursail}
\alias{foursail}
\title{SAIL model}
\usage{
foursail(refl, tran, rsoil, param)
}
\arguments{
\item{refl}{input leaf reflectance from 400-2500nm (can be measured or modeled)}

\item{tran}{input leaf transmittance from 400-2500nm (can be measured or modeled)}

\item{rsoil}{input soil reflectance spectra from 400-2500nm (can be measured or modeled)}

\item{param}{Vector of SAIL parameter values:
* LIDFa: Leaf angle distribution function - parameter a
* LIDFb: Leaf angle distribution function - parameter b
* TypeLIDF: Leaf angle distribution function type (1 or 2)
* LAI: Leaf area index
* q: Hot spot effect parameter
* tts: Solar zenith angle
* tto: Observer zenith angle
* psi: Sun-sensor azimuth angle}
}
\value{
Spectra matrix (see \code{\link[=spectra]{spectra()}}) (2101 x 4) of reflectance factors
for wavelengths 400 to 2500nm:
* bi-hemispherical reflectance (rddt)
* hemispherical directional (rsdt)
* directional hemispherical (rdot)
* bi-directional (rsot)
}
\description{
R wrapper for 4SAIL model
}
\author{
Shawn Serbin

Alexey Shiklomanov
}
