% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_plate_model.R
\name{generalized_plate_model}
\alias{generalized_plate_model}
\title{Generalized plate model}
\usage{
generalized_plate_model(k, refractive, N)
}
\arguments{
\item{k}{Specific absorption coefficient (400 - 2500nm). See also
\code{\link[=dataspec_4]{dataspec_4()}} and friends.}

\item{refractive}{Refractive index (400 - 2500nm). See
\code{\link[=refractive_45]{refractive_45()}} and friends.}

\item{N}{Effective number of mesophyll layers (see \code{\link[=prospect]{prospect()}})}
}
\description{
This is the fundamental physical model underlying the PROSPECT family of leaf RTMs.
}
