% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate-rsr.R
\name{generate.rsr.all}
\alias{generate.rsr.all}
\title{Generate RSR matrices for all sensors and return as list}
\usage{
generate.rsr.all()
}
\description{
Only needs to be called when updating these funcitons with new
data
}

