% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm.R
\name{gpm}
\alias{gpm}
\title{Generalized plate model}
\usage{
gpm(k, refractive, N)
}
\arguments{
\item{k}{Specific absorption coefficient (400 - 2500nm)}

\item{refractive}{Refractive index (400 - 2500nm)}

\item{N}{Effective number of mesophyll layers (see \code{\link[=prospect]{prospect()}})}
}
\description{
This is the fundamental physical model underlying the PROSPECT family of leaf RTMs.
}
