% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.auto.R
\name{invert.auto}
\alias{invert.auto}
\title{Inversion with automatic convergence checking}
\usage{
invert.auto(observed, invert.options, return.samples = TRUE,
  save.samples = NULL, quiet = FALSE, parallel = TRUE,
  parallel.cores = NULL)
}
\arguments{
\item{observed}{Matrix of observed values. Must line up with output of 'model'.}

\item{invert.options}{R list object containing the following elements:

inits Vector of initial values of model parameters to be inverted.

ngibbs Number of MCMC iterations

prior.function Function for use as prior. Should take a vector of parameters 
as input and return a single value -- the sum of their log-densities -- as 
output.

param.mins Vector of minimum values for inversion parameters

model The model to be inverted. This should be an R function that takes 
`params` as input and returns one column of `observed` (nrows should be the 
same). Constants should be implicitly included here.

adapt Number of steps for adapting covariance matrix (i.e. adapt every 'n' 
steps). Default=100

adj_min Minimum threshold for rescaling Jump standard deviation.  Default = 
0.1.

target Target acceptance rate. Default=0.234, based on recommendation for 
multivariate block sampling in Haario et al. 2001

do.lsq Perform least squares optimization first (see `invert.lsq`), and use 
outputs to initialize Metropolis Hastings. This may improve mixing time, but 
risks getting caught in a local minimum.  Default=FALSE

nchains Number of independent chains.

inits.function Function for randomly generating initial conditions.

ngibbs.max Maximum number of total iterations (per chain). DEFAULT = 5e6

ngibbs.min Minimum number of total iterations (per chain). DEFAULT = 5000.

ngibbs.step Number of iterations between convergence checks. Default = 1000.}

\item{return.samples}{Include full samples list in output. Default = TRUE.}

\item{save.samples}{Save samples to file as the inversion proceeds (useful for debugging).
If NULL, do not save samples. Default = NULL.}

\item{parallel}{Logical. Whether or not to run multiple chains in parallel on multiple cores (default = TRUE).}

\item{parallel.cores}{Number of cores to use for parallelization. If NULL (default), allocate one fewer than detected number of cores.}
}
\value{
List of "results" (summary statistics and Gelman Diagnostic) and "samples" (mcmc.list object, or "NA" if return.samples=FALSE)
}
\details{
Performs an inversion via the `invert.custom` function with multiple chains and automatic convergence checking. 
Convergence checks are performed using the multivariate Gelman-Rubin diagnostic.
}

