% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.fast.R
\name{invert.fast}
\alias{invert.fast}
\title{Bayesian RTM inversion: Fortran implementation}
\usage{
invert.fast(modname, observed, inits, cons, pmu, psd, plog, minp, ngibbs)
}
\arguments{
\item{modname}{Name of the model to invert (character). 
Refer to `data/model.list.csv`}

\item{observed}{Observed reflectance. Can be a vector, matrix, 
or data.frame; BUT NOTE: all are coerced via "as.matrix"}

\item{inits}{Named vector of parameters to invert. Names are required!}

\item{cons}{Named vector of constants. Names are required!}

\item{pmu}{Numeric vector of prior means for inversion parameters. 
Must line up with `inits`}

\item{psd}{Numeric vector of prior standard deviations for inversion 
parameters. Must line up with `inits`.}

\item{plog}{Logical vector. Whether or not to use lognormal 
distribution for the prior. NOTE: `pmu` and `psd` are 
the distribution parameters, NOT the distribution's actual 
mean and standard deviation.}

\item{minp}{Numeric vector of minimum values for parameters.}

\item{ngibbs}{Number of iterations for MCMC}
}
\value{
Matrix (ngibbs x (length(inits) + 1)) of MCMC samples 
         of parameters.
}
\details{
Performs a Bayesian inversion of a Radiative 
         Transfer Model. Sampling is performed using an 
         adaptive Metropolis-Hastings algorithm operating 
         independently on each parameter. The model takes 
         both paramters and constants as an argument -- 
         parameters (`inits`) are sampled and treated as 
         random (unknown) while constants (`cons`) are fixed. 
         Normal or log-normal priors are be specified by 
         passing a vector of means (`pmu`), standard 
         deviations (`psd`), and a logical of whether or not 
         to use the lognorml (`plog`). Minimum values for 
         the parameters must be given as a vector (`minp`).
}
\author{
Alexey Shiklomanov
}

