% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invert.fast.re.R
\name{invert.fast.re}
\alias{invert.fast.re}
\title{Bayesian inversion with random effects}
\usage{
invert.fast.re(modname, observed, inits, rand, cons, pmu, psd, plog, minp,
  ngibbs)
}
\arguments{
\item{modname}{Name of the model to invert (character).
Refer to `model.list`}

\item{observed}{Observed reflectance. Can be a vector,
matrix, or data.frame; BUT NOTE: all are coerced to
matrix via `as.matrix`.}

\item{inits}{Named vector of parameters to invert. Names
are required!}

\item{rand}{Numeric matrix of initial values for random
effects, with dimensions (npars x nspec), where npars is
the number of parameters and nspec is the number of
spectra.}

\item{cons}{Numeric vector of constants. Names are
required!}

\item{pmu}{Numeric vector of prior means for inversion
parameters. Must line up with `inits`}

\item{psd}{Numeric vector of prior standard deviations
for inversion parameters. Must line up with `inits`.}

\item{plog}{Logical vector. Whether or not to use
lognormal distribution for the prior. NOTE: `pmu` and `psd`
are the distribution parameters, NOT the distribution's
actual mean and standard deviation.}

\item{minp}{Numeric vector of minimum values for parameters.}

\item{ngibbs}{Number of iterations for MCMC}
}
\value{
Matrix (ngibbs x (npars*(nspec+2)+1)) of MCMC
samples of parameters.
}
\description{
Bayesian inversion with random effects
}
\details{
Performs a Bayesian inversion of a Radiative
Transfer model with individual random effects. Sampling
is performed using an adaptive Metropolis-Hastings
algorithm operating independently on each parameter. See
also: `invert.fast`.
}
\author{
Alexey Shiklomanov
}

