% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestools.R
\name{invert_bt}
\alias{invert_bt}
\title{Perform Bayesian inversion using BayesianTools package}
\usage{
invert_bt(observed, model, prior, custom_settings = list())
}
\description{
Perform Bayesian inversion using BayesianTools package
}
