% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invert_prospect_re}
\alias{invert_prospect_re}
\title{Bayesian inversion of PROSPECT4 model with random effects}
\usage{
invert_prospect_re(spectra, ngibbs = 15000, adapt = 25, min_adapt = 0.05)
}
\arguments{
\item{spectra}{A matrix of reflectance spectra with dimensions 2101 x n,
where n is the number of spectra. Vectors are coerced to matrices.}

\item{ngibbs}{The number of MCMC steps for the iteration. Default is 15000.
Convergence for good, individual spectra happens around 5000 steps.}

\item{adapt}{How often to adapt the Jump distribution, in MCMC steps. Default = 25}

\item{min_adapt}{Minimum value to adapt by. This prevents the Jump distribution variance
from going to zero when no values are accepted. Default is 0.05.}
}
\value{
Matrix of samples of each PROSPECT parameter, residual variance, and
    random effects variance for each parameter, and random effects values
    for each leaf.
}
\description{
Bayesian inversion of PROSPECT4 model with random effects
}
\examples{
data(testspec)
samples <- invert_prospect_re(testspec_ACRU[,1:5], 1000)
dim(samples)
par(mfrow=c(2,2))
plot(samples[,1], type='l', ylab="N")
plot(samples[,2], type='l', ylab="Cab")
plot(samples[,3], type='l', ylab="Cw")
plot(samples[,4], type='l', ylab="Cm")
}
\author{
Alexey Shiklomanov
}

