% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{lognorm.mu}
\alias{lognorm.mu}
\title{Functions for default priors
Lognormal mean parameters}
\usage{
lognorm.mu(mean, sd)
}
\arguments{
\item{mean}{Sample mean}

\item{sd}{Sample standard deviation}
}
\description{
Functions for default priors
Lognormal mean parameters
}
