% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{lognorm.sigma}
\alias{lognorm.sigma}
\title{Lognormal sigma parameter}
\usage{
lognorm.sigma(mean, sd)
}
\arguments{
\item{mean}{Sample mean}

\item{sd}{Sample standard deviation}
}
\description{
Lognormal sigma parameter
}
