% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matplot.spectra.R
\name{matplot.default}
\alias{matplot.default}
\title{Matplot default method}
\usage{
\method{matplot}{default}(...)
}
\arguments{
\item{...}{arguments passed to \code{matplot}}
}
\description{
Matplot default method
}
