% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neff.R
\name{neff}
\alias{neff}
\title{Effective sample size}
\usage{
neff(x, ...)
}
\arguments{
\item{x}{A vector or time series}

\item{...}{additional arguments passed to methods}
}
\description{
Calculate effective sample size of vector based on its autocorrelation.
}
