% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\docType{data}
\name{params.prospect4}
\alias{params.prospect4}
\title{PROSPECT 4 parameters}
\format{
An object of class \code{character} of length 4.
}
\usage{
params.prospect4
}
\description{
Shortcut lists for PROSPECT parameter names
}
\keyword{datasets}
