% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\docType{data}
\name{params.prospect5}
\alias{params.prospect5}
\title{PROSPECT 5 parameters}
\format{
An object of class \code{character} of length 5.
}
\usage{
params.prospect5
}
\description{
PROSPECT 5 parameters
}
\keyword{datasets}
