% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\docType{data}
\name{params.prospect5b}
\alias{params.prospect5b}
\title{PROSPECT 5B parameters}
\format{
An object of class \code{character} of length 6.
}
\usage{
params.prospect5b
}
\description{
PROSPECT 5B parameters
}
\keyword{datasets}
