% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\docType{data}
\name{params.prospectd}
\alias{params.prospectd}
\title{PROSPECT D parameters}
\format{
An object of class \code{character} of length 7.
}
\usage{
params.prospectd
}
\description{
PROSPECT D parameters
}
\keyword{datasets}
