% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params2edr.R
\name{params2edr}
\alias{params2edr}
\title{Convert named parameter vector to EDR-compatible inputs}
\usage{
params2edr(params, sep = ".", prospect = TRUE, version = 5)
}
\arguments{
\item{params}{Named parameter vector}

\item{sep}{Separator between PFT name and trait name. Must be a single
character (default = ".").}

\item{prospect}{Logical. If \code{TRUE} (default), scan for PROSPECT traits and
pass them to PROSPECT.}

\item{version}{PROSPECT version}
}
\value{
List containing \code{spectra_list} and \code{trait.values}, both objects needed by \link{EDR}.
}
\description{
Creates a nested list whose components are suitable for passing to EDR.
}
\details{
If \code{prospect = TRUE}, parameters prefixed with \code{prospect_} are passed to
\link{prospect} with the specified \code{version}, and other parameters are passed to
\code{trait.values}.

The regular expression defining the separation is greedy, i.e.
\code{temperate.Early_Hardwood.SLA} will separate into \code{temperate.Early_Hardwood}
and \code{SLA} (assuming the default \code{sep = "."}). Therefore, it is crucial that
trait names not contain any \code{sep} characters (neither ED nor PROSPECT
parameters should anyway). If this is a problem, use an alternate separator
(e.g. \code{|}).

Note that using \code{sep = "."} allows this function to directly invert the
default behavior of \code{unlist}. That is, calling
\code{unlist(params2edr(params, prospect = FALSE)$trait.values)} will return the input vector of
trait values. This makes \code{unlist} a convenient way to go from a
\code{trait.values} list to a properly formatted \code{params} vector.

Because unused ED parameters in the \code{config.xml} are ignored, the PROSPECT
parameters are saved in the \code{trait.values} object as well, which may be
useful for debugging.
}
\author{
Alexey Shiklomanov
}
