\name{plot.prospect.inv}
\alias{plot.prospect.inv}
\title{Diagnostic plots displaying the results of the PROSPECT model inversion}
\usage{
  plot.prospect.inv(inv.output, outdir, file)
}
\arguments{
  \item{inv.output}{output from the PROSPECT inversion}

  \item{outdir}{output directory for diagnostic plots}

  \item{file}{file name for diagnostic plots}
}
\value{
  plot diagnostic plots of model inversion
}
\description{
  Diagnostic plots displaying the results of the PROSPECT
  model inversion
}
\author{
  Shawn Serbin
}

