% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spectra.R
\name{plot.spectra}
\alias{plot.spectra}
\title{Plot spectra vs. wavelength}
\usage{
\method{plot}{spectra}(x, y = NULL, type = "l", ...)
}
\arguments{
\item{x}{A spectra object}

\item{y}{Ignored.}

\item{type}{plot style, e.g. "l" for lines, "b" for lines and points}

\item{...}{Additional arguments to \code{plot}}
}
\description{
Plot spectra vs. wavelength
}
