% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spectra.R
\name{print.spectra}
\alias{print.spectra}
\title{Print method for spectra S3 class}
\usage{
print.spectra(x, n = 10, ...)
}
\arguments{
\item{x}{A spectra object}

\item{n}{Max number of rows to print (show first \code{n/2} and last \code{n/2} rows)}

\item{...}{Additional arguments to \code{print}}
}
\description{
Print method for spectra S3 class
}
