% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results_summary.R
\name{print_results_summary}
\alias{print_results_summary}
\title{Neatly print inversion results summary}
\usage{
print_results_summary(output)
}
\arguments{
\item{output}{Output from \code{invert.auto}}
}
\description{
Neatly print inversion results summary
}
\author{
Alexey Shiklomanov
}
