% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior.R
\name{prior.defaultvals.prospect}
\alias{prior.defaultvals.prospect}
\title{Default prior parameters for PROSPECT models}
\usage{
prior.defaultvals.prospect(sd.inflate = 3)
}
\arguments{
\item{sd.inflate}{Standard deviation multiplier (default = 3)}
}
\description{
Default prior parameters for PROSPECT models
}

