% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{priorfunc.prospect}
\alias{priorfunc.prospect}
\title{Default PROSPECT 5 prior function}
\usage{
priorfunc.prospect(pmu, psigma)
}
\arguments{
\item{pmu}{Lognormal mu parameter}

\item{psigma}{Lognormal sigma parameter}
}
\details{
Assumes lognormal distribution for all parameters. NOTE that prior 
on N is shifted by 1.
}

