% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro2s.R
\name{pro2s}
\alias{pro2s}
\title{Coupled PROSPECT-Two-stream model}
\usage{
pro2s(param, prospect.version = 5)
}
\arguments{
\item{param}{Model parameters, in the following order: N, Cab, (Car,
Cbrown), Cw, Cm, solar zenith angle, LAI, soil_moisture}

\item{prospect.version}{Version of PROSPECT to use (4, 5, or '5B'; default=5)}
}
\value{
Spectra matrix (see \code{\link[=spectra]{spectra()}}) for wavelengths 400 to 2500nm containing the following columns:
\itemize{
\item \code{alpha.c} -- Direct ("collimated") reflectance
\item \code{alpha.i} -- Diffuse ("isotropic") reflectance
\item \code{Tc} -- Direct transmittance to background
\item \code{Ti} -- Diffuse reflectance to background
\item \code{Ac} -- Direct absorbance by canopy
\item \code{Ai} -- Diffuse absorbance by canopy
}
}
\description{
Coupled PROSPECT-Two-stream model
}
