% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosail.R
\name{pro4sail}
\alias{pro4sail}
\title{PRO4SAIL model}
\usage{
pro4sail(param)
}
\arguments{
\item{param}{Vector of PRO4SAIL parameter values:
N: Effective number of leaf layers (>1)
Cab: Leaf chlorophyll content (ug/cm2) (>0)
Car: Leaf carotenoid content (ug/cm2) (>0)
Cbrown: Leaf brown matter content (ug/cm2) (>0)
Cw: Leaf water content (cm) (>0)
Cm: Leaf dry matter content (ug/cm2) (>0)
LIDFa: Leaf angle distribution function - parameter a
LIDFb: Leaf angle distribution function - parameter b
TypeLIDF: Leaf angle distribution function type (1 or 2)
LAI: Leaf area index
q: Hot spot effect parameter
tts: Solar zenith angle
tto: Observer zenith angle
psi: Sun-sensor azimuth angle
psoil: Fraction of soil moisture}
}
\value{
Matrix (2101 x 4) of reflectance factors:
     rddt: bidirectional
     rsdt: hemispherical directional
     rdot: directional hemispherical
     rsot: bi-directional
}
\details{
R wrapper for PRO4SAIL model
}
\author{
Alexey Shiklomanov
}

