\name{prospect4}
\alias{prospect4}
\title{PROSPECT-4 leaf radiative transfer model}
\usage{
  prospect4(N, Cab, Cw, Cm)
}
\arguments{
  \item{N}{leaf structure parameter}

  \item{Cab}{chlorophyll a+b content in ug/cm2}

  \item{Cw}{equivalent water thickness (EWT) in g/cm2 or
  cm-1}

  \item{Cm}{dry matter content in g/cm2 (alias LMA)}
}
\description{
  Plant leaf reflectance and transmittance are calculated
  from 400 nm to 2500 nm (1 nm step) with the following
  parameters:
}
\author{
  Shawn Serbin
}
\references{
  Jacquemoud S., Ustin S.L., Verdebout J., Schmuck G.,
  Andreoli G., Hosgood B. (1996), Estimating leaf
  biochemistry using the PROSPECT leaf optical properties
  model, Remote Sens. Environ., 56:194-202.

  Jacquemoud S., Baret F. (1990), PROSPECT: a model of leaf
  optical properties spectra, Remote Sens. Environ.,
  34:75-91.

  Feret et al. (2008), PROSPECT-4 and 5: Advances in the
  Leaf Optical Properties Model Separating Photosynthetic
  Pigments, Remote Sensing of Environment

  Feret et al. (2008).
  http://teledetection.ipgp.jussieu.fr/prosail/
}

