% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{resample.default}
\alias{resample.matrix}
\alias{resample.spectra}
\title{Resample vector, matrix, or spectra}
\usage{
resample(values, ...)

\method{resample}{default}(values, from, to, method = "fmm", ...)

\method{resample}{matrix}(values, from, to, ...)

\method{resample}{spectra}(values, to, ...)
}
\arguments{
\item{values}{Vector or matrix of values, or object of class \code{\link[=spectra]{spectra()}}.
Length of vector, or \code{nrow} of matrix must match length of \code{from}. For
\code{spectra}, \code{from} argument is omitted because it is taken from
\code{wavelengths}.}

\item{...}{Additional arguments to \code{\link[stats:splinefun]{stats::splinefun()}} or
\code{\link[stats:approxfun]{stats::approxfun()}}}

\item{from}{X values for interpolation (for \code{spectra} objects, this is
assumed to be the \code{wavelengths} attribute.)}

\item{to}{Y values onto which to interpolate. For \code{spectra} objects, this
should be new wavelengths.}

\item{method}{One of the methods for \code{\link[stats:splinefun]{stats::splinefun()}} (for polynomial
and periodic splines) or \code{\link[stats:approxfun]{stats::approxfun()}} (for constant or linear).
Default is \code{"fmm"} (same as splinefun).}
}
\value{
Object of the same class as \code{values}, resampled to the \code{to} values.
}
\description{
Convenient wrapper around base R's \code{splinefun} and \code{approxfun}. See
\code{\link[stats:splinefun]{stats::splinefun()}} and \code{\link[stats:approxfun]{stats::approxfun()}} documentation for information
on different spline methods and additional arguments.
}
