% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rsr.R
\name{rsr.from.fwhm}
\alias{rsr.from.fwhm}
\title{Generate relative spectral response (RSR) matrix based on FWHM data}
\usage{
rsr.from.fwhm(wavelength, fwhm)
}
\arguments{
\item{wavelength}{Vector of average band widths, as reported in FWHM data.}

\item{fwhm}{Vector of full-width half maximum (FWHM) bandwidths, as reported 
in FWHM data.}
}

