\name{rtm-package}
\alias{rtm-package}
\alias{rtm}
\docType{package}
\title{
	Relational Topic Models
}
\description{
This package implements the Relational Topic Model which models document
collections connected by links.  Learned models can be used to make
predictions about content and connectivity.  For efficiency, this model
also uses the SparseLDA optimizations described by [Mimno].
}
\details{
\tabular{ll}{
Package: \tab rtm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-10-29\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
  See demos.
}
\author{
Jonathan Chang <jonchang@fb.com>
}
\references{
	LDA
	RTM
	SparseLDA
}
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
