\name{RTM-package}
\alias{RTM}
\title{Reaction Transport Modelling Course - tutorials}
\description{
This package contains a  number of tutorials, that serve in preparation for the exam for the course taught at Utrecht University, entitled:

Reaction Transport Modelling in the Hydrosphere.

Tutors: Lubos Polerecky and Karline Soetaert

To toggle the tutorials, just write:

run_tutorial(name_of_tutorial, package = "RTM")

or use function \link{RTMtutorial}
}

\examples{
# To have a list of all available tutorials:

RTMtutorial("?")
}

