% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestools_helpers.R
\name{rtm_loglike}
\alias{rtm_loglike}
\title{Model-error normal log-likelihood generator}
\usage{
rtm_loglike(nparams, model, observed, sweep = FALSE, verbose = TRUE,
  heteroskedastic = FALSE, ...)
}
\arguments{
\item{nparams}{Number of model parameters. Note that this is 1 or 2 fewer
(depending on the error model) than the number of parameters passed to the
sampler.}

\item{model}{Model function to call, whose outputs must line up with
observations.}

\item{observed}{Observation vector or matrix. Must line up with dimensions
of \code{model} output}

\item{sweep}{Logical. If \code{TRUE}, use \code{sweep} to calculate the model-dat
difference, which allows dimension mismatch but is slower. Otherwise, use
\code{+}.}

\item{verbose}{Logical. If \code{TRUE}, print verbose error messages when
likelihood evaluation fails.}

\item{heteroskedastic}{Logical. If \code{TRUE}, residual SD is a linear function
of the model output (note that this requires 2 residual parameters). If
\code{FALSE} (default), it's a constant.}
}
\value{
Log-likelihood function of a parameter vector, which can be used in
a \code{bayesianSetup}
}
\description{
Model-error normal log-likelihood generator
}
