% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncnorm.R
\name{rtnorm}
\alias{rtnorm}
\title{Random sampling from one-sided truncated normal distribution}
\usage{
rtnorm(mu, sd, MIN)
}
\arguments{
\item{mu}{The mean parameter of the truncated normal. NOTE that this is NOT
equal to the mean of the distribution}

\item{sd}{The standard deviation parameter of the truncated normal. Again, NOTE
tht this is not the SD of the distribution.}

\item{MIN}{The minimum value, which defines the truncation.}
}
\value{
Numeric length one.
}
\description{
Truncated normal distribution functions
}
\details{
Draws a random number and, if it doesn't fall within the 
     specified range, resample using an adjusted Normal CDF. This isn't
     performed immediately because CDF sampling calls three functions --
     qnorm, runif, and pnorm--and therefore is much less efficient than a
     simple random sample.
}
\author{
Alexey Shiklomanov
}

