% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_until_converged.R
\name{sample_until_converged}
\alias{sample_until_converged}
\title{Run BayesianTools sampler until convergence has been achieved}
\usage{
sample_until_converged(bt_input, common_settings = list(),
  first_settings = list(), loop_settings = list(), sampler = "DEzs",
  max_iter = 1e+06, min_samples = 5000, use_mpsrf = FALSE,
  threshold = 1.1, save_progress = NULL)
}
\arguments{
\item{bt_input}{Either previous BayesianTools output (class \code{mcmcsampler})
or a BayesianSetup object}

\item{common_settings}{BayesianTools settings list}

\item{first_settings}{Settings for first inversion attemp}

\item{loop_settings}{Settings for subsequent inversion attempts}

\item{sampler}{Character, which sampling algorithm to use. Default is "DEzs"}

\item{max_iter}{Maximum total number of iterations. If exceeded, return
samples even if not converged. Default = 1e6.}

\item{min_samples}{Minimum number of samples to return after burnin. Even if
converged, sampling will resume until this number is met. Default = 5000.}

\item{use_mpsrf}{Use multivariate PSRF in burnin calculations. Default = FALSE.}

\item{threshold}{PSRF convergence threshold. Default = 1.1}

\item{save_progress}{Target RDS file for saving samples between sampling
steps. If \code{NULL} (default), don't save samples during runs.}
}
\value{
BayesianTools \code{mcmcSampler} object
}
\description{
Run BayesianTools sampler until convergence has been achieved
}
