% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-response.R
\docType{data}
\name{sensor.proper}
\alias{sensor.proper}
\title{Sensor list with proper names}
\format{
An object of class \code{character} of length 12.
}
\usage{
sensor.proper
}
\description{
Sensor list with proper names
}
\keyword{datasets}
