% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectra_types}
\alias{spectra_types}
\title{Retrieve spectra type of spectra object}
\usage{
spectra_types(spectra)
}
\arguments{
\item{spectra}{Object of class \code{spectra}}
}
\description{
Retrieve spectra type of spectra object
}
