% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectral-response.R
\name{spectral.response}
\alias{spectral.response}
\title{Convolution of spectra to sensor RSR}
\usage{
spectral.response(spec, sensor)
}
\arguments{
\item{spec}{Full (1 nm) spectrum (vector)}

\item{sensor}{Sensor name (string). See sensor.list}
}
\description{
Convolution of spectra to sensor RSR
}

