% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.spectra.R
\name{[.spectra}
\alias{[.spectra}
\title{Select spectra}
\usage{
\method{[}{spectra}(spectra, i, j, drop = FALSE)
}
\arguments{
\item{spectra}{Object of class \code{spectra}}

\item{i, j}{indices specifying elements to extract or replace.
See \code{\link[base:Extract]{base::Extract}} for details.}

\item{drop}{Coerce result to the lowest dimension possible?
See \code{\link[base:drop]{base::drop()}} for details.}
}
\description{
Select spectra
}
