% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.spectra.R
\name{[[.spectra}
\alias{[[.spectra}
\title{Select spectra by wavelength}
\usage{
\method{[[}{spectra}(spectra, wavelength, j)
}
\arguments{
\item{spectra}{Object of class \code{spectra}}

\item{wavelength}{Wavelength vector to select}

\item{j}{index specifying elements to extract or replace.
See \code{\link[base:Extract]{base::Extract}} for details.}
}
\description{
Select spectra by wavelength
}
