% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.spectra.R
\name{[[<-.spectra}
\alias{[[<-.spectra}
\title{Assign values to spectra by wavelength}
\usage{
\method{[[}{spectra}(spectra, wavelength, j, values) <- value
}
\description{
Assign values to spectra by wavelength
}
