% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-output.R
\name{summary_simple}
\alias{summary_simple}
\title{Simple summary statistics on MCMC samples}
\usage{
summary_simple(samples)
}
\arguments{
\item{samples}{Matrix of MCMC samples}
}
\description{
Calculate simple univariate summary statistics and return as named list
}
