\name{tav}
\alias{tav}
\title{Transmission of radiation through elementary layers within PROSPECT}
\usage{
  tav(teta, ref)
}
\arguments{
  \item{teta}{transmission}

  \item{ref}{reflectance}
}
\description{
  Transmission of isotropic radiation across an interface
  between two dielectrics The transmission of radiation
  through elementary layers within PROSPECT
}
\author{
  Shawn Serbin
}
\references{
  Stern F. (1964), Transmission of isotropic radiation
  across an interface between two dielectrics, Appl. Opt.,
  3(1):111-113.

  Allen W.A. (1973), Transmission of isotropic light across
  a dielectric surface in two and three dimensions, J. Opt.
  Soc. Am., 63(6):664-666.

  Feret et al. (2008), PROSPECT-4 and 5: Advances in the
  Leaf Optical Properties Model Separating Photosynthetic
  Pigments, Remote Sensing of Environment

  Feret et al. (2008).
  http://teledetection.ipgp.jussieu.fr/prosail/
}

