% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{tidy_spectra}
\alias{tidy_spectra}
\title{Tidy spectra}
\usage{
tidy_spectra(spectra)
}
\arguments{
\item{spectra}{Object of class \code{spectra}.}
}
\value{
Long \code{tbl_df}. See Details.
}
\description{
Convert spectra to a long tidy \code{tbl_df}.
}
\details{
Return object has the following columns.
\itemize{
\item \code{col_name} -- Matches the column name of the original spectra
\item \code{col_rep} -- The count of a column name in the spectra, useful for
identifying repeat column names.
\item \code{spectra_type} -- Type of spectra
\item \code{wavelength} -- Wavelength, in nm
\item \code{value} -- The spectra value
}
}
