% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rsr.R
\name{trim.rsr}
\alias{trim.rsr}
\title{Trim RSR matrix to wavelength limits}
\usage{
trim.rsr(rsr, wl.min = 400, wl.max = 2500)
}
\arguments{
\item{rsr}{RSR matrix}

\item{wl.min}{Minimum wavelength (inclusive, default = 400)}

\item{wl.max}{Maximum wavelength (inclusive, default = 2500)}
}

