% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{vec_reps}
\alias{vec_reps}
\title{Index repeated elements of a vector}
\usage{
vec_reps(x)
}
\arguments{
\item{x}{Vector whose contents should be indexed}
}
\value{
Vector of integer indices equal to length of \code{x}
}
\description{
For example, vector \code{a,a,b,b,a,c} would have indices \code{1,2,1,2,3,1}.
}
