% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{wavelengths}
\alias{wavelengths}
\title{Retrieve wavelengths from spectra object}
\usage{
wavelengths(spectra)
}
\arguments{
\item{spectra}{Object of class \code{spectra}}
}
\description{
Retrieve wavelengths from spectra object
}
