% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvt.R
\name{bvt}
\alias{bvt}
\alias{rbvt}
\alias{dbvt}
\title{Bivariate \eqn{t}-distributions}
\usage{
rbvt(n, df, theta)

dbvt(t1, t2, df, theta, cutoff, truncation = NULL)
}
\arguments{
\item{n}{the number of samples required}

\item{df}{a numeric vector giving the degree of freedoms of variables}

\item{theta}{a numeric value of mixing angle}

\item{t1}{a numeric vector of baseline measurements}

\item{t2}{a numeric vector of followup measurements}

\item{cutoff}{a numeric value of cutoff point in which the baseline measurements are above or below}

\item{truncation}{a character string; one of \code{"left"} or \code{"right"} indicating the specification of distribution.}
}
\value{
It give the matrix of random numbers and density whose dimension is determine by \code{n}, \eqn{t_1}, and \eqn{t_2}.
}
\description{
Generating random numbers, and density from bivariate \eqn{t}-distributions either truncated or complete.
The mean of complete distribution is equal to zero
}
\details{
The bivariate \eqn{t}-distribution have identical marginals and equal or unequal degree of freedoms proposed by Shaw, W. T., & Lee, K. T. A. (2008).
The density function of bivariate \eqn{t}-distribution with equal degree of freedom is
\deqn{f\left( {t_1},{t_2};n,\theta \right) = \frac{1}{{2\pi \cos \left( \theta  \right)}}\frac{1}{{{{\left( {1 + \frac{\Delta }{n}} \right)}^{n/2 + 1}}}}}
where \eqn{-\infty < {t_1} < \infty}, \eqn{- \infty  < {t_2} < \infty}, \eqn{\Delta  = ({t_1^2 + t_2^2 - 2{t_1}{t_2}\sin (\theta )})^{-1}{{{\cos }^2}(\theta )}}, \eqn{n} is the degree of freedom, \eqn{\theta} is the mixing angle.
}
\examples{
  ### Generating a 1,000 samples
randomT<-rbvt(n=1000,df=c(15,9),theta=pi/4)
hist(randomT[,1])

  #create bivariate t-distribution
t1    <- sort(randomT[,1])
t2    <- sort(randomT[,2])
df    <- c(15,9)
f     <- function(t1, t2) dbvt(t1,t2,df=c(15,9),theta= pi/4)
z     <- outer(t1, t2, f)

  #create contour plot
contour(t1, t2, z)

  ### Compute density from truncated distribution
randTrunc<-randomT[randomT[,1]>=0.5,]
  ### create right trucatated bivariate t-distribution
t1.Tr <- sort(randTrunc[,1])
t2.Tr <- sort(randTrunc[,2])
df    <- c(15,9)
f     <- function(t1.Tr, t2.Tr) dbvt(t1.Tr,t2.Tr,
df=c(15,9),theta= pi/4,cutoff = 0.5,truncation = "right")
z     <- outer(t1.Tr, t2.Tr, f)

  #create contour plot
contour(t1.Tr, t2.Tr, z)
}
\references{
Shaw, W. T., & Lee, K. T. A. (2008). Bivariate Student t distributions
with variable marginal degrees of freedom and independence. \emph{Journal of Multivariate Analysis}, 99(6), 1276-1287.
}
\seealso{
\code{\link{RTM.T}}
}
\author{
Shaw, W. T., & Lee, K. T. A.
}
