% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defparam.R
\name{defparam}
\alias{defparam}
\title{Get default parameters}
\usage{
defparam(modname)
}
\arguments{
\item{modname}{Model name. Must match `modname` in `model.list`}
}
\value{
Named vector of default parameter values
}
\details{
Extract default parameter values from `model.list`
}

