% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.wrapper.R
\name{EDR}
\alias{EDR}
\title{ED radiative transfer module (EDR) wrapper function}
\usage{
EDR(img_path, ed2in_path, spectra_list, trait.values,
  soil_reflect_path = system.file("extdata", "soil_reflect_par.dat",
  package = "PEcAnRTM"), wood_reflect_path = system.file("extdata",
  "wood_reflect_par.dat", package = "PEcAnRTM"), par.wl = 400:2499,
  nir.wl = 2500, edr_exe_path = NULL,
  output.path = dirname(normalizePath(ed2in_path, mustWork = TRUE)),
  settings = list(model = list(revision = "git", config.header = NULL)),
  singularity_args = list(), clean = FALSE, stderr = TRUE,
  verbose_error = TRUE, ...)
}
\arguments{
\item{img_path}{Path to Singularity container (usually a \code{.simg} file)}

\item{ed2in_path}{Path to ED2IN file.}

\item{spectra_list}{List of spectral data matrices. Names must exactly match
the PFTs given in \code{trait.values}. Each item must be a matrix of
reflectance (labeled 'R' or 'reflectance') and transmittance (labeled 'T' or
'transmittance') values. If the matrix is not of class \code{spectra} (see
\code{\link[=spectra]{spectra()}}), it must also have a column of wavelength values (labeled
'wl'). Such a matrix is returned by default by all versions of PROSPECT in
this package.}

\item{trait.values}{Named, hierarchical list of trait values for generating
config.xml file. Names must be PFTs, and must exactly match names of
\code{spectra_list}.}

\item{soil_reflect_path}{Path to soil reflectance file (defaults to spectra
in package \code{extdata})}

\item{wood_reflect_path}{Path to wood reflectance file (defaults to spectra
in package \code{extdata})}

\item{par.wl}{Vector of wavelengths defining PAR region}

\item{nir.wl}{Vector of wavelengths defining NIR region}

\item{edr_exe_path}{If \code{img_path} is \code{NULL}, a full path to the EDR
executable. Ignored otherwise.}

\item{output.path}{Directory in which to execute the run. Defaults to
\code{dirname(ed2in_path)}.}

\item{settings}{PEcAn settings list. Default is \code{list(model = list(revision = "git", config.header = NULL))}.}

\item{singularity_args}{Additional arguments to be passed to \code{singularity run} (before)}

\item{clean}{Logical. If \code{TRUE}, remove all files generated by this function
(e.g. cloned history file, ED2IN, output HDF files).}

\item{stderr}{Logical. If \code{TRUE} (default), internalize \code{system2} results as
R character vector. \code{TRUE} is recommended because it allows EDR to check its
execution and to run more quietly.}

\item{verbose_error}{Logical. If \code{TRUE} (default), spit out the full ED
if EDR execution fails.}

\item{...}{Additional arguments to \code{system2}}
}
\description{
This function provides a convenient way to call the ED
radiative transfer module (EDR, which simulates full spectral return of an
ED patch for a given point in time) directly from R.
}
\author{
Alexey Shiklomanov
}
