% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.wrapper.R
\name{EDR}
\alias{EDR}
\title{ED radiative transfer module (EDR) wrapper function}
\usage{
EDR(paths, RT.matrix, par.wl, nir.wl, datetime, history.prefix = "history",
  edr.exe.name = "ed_2.1-opt", change.history.time = TRUE,
  output.path = getwd(), clean = FALSE)
}
\arguments{
\item{paths}{List of relevant paths. Must contain the following:
`ed2in` -- Path to ED2IN file for the original ED run. If this is `NA`, the 
function assumes that the ED2IN file in `output.path` is already up-to-date 
and skips the analysis step.
`history` -- Path and prefix for history file for the run of interest;
`edr.exe` -- Path to EDR executable}

\item{RT.matrix}{Matrix of wavelengths, reflectance, and transmittance 
values. Matrix must have column names "wl" (wavelength), "R" (reflectance), 
and "T" (transmittance). Such a matrix is returned by default by all 
versions of PROSPECT in this package.}

\item{par.wl}{Vector of wavelengths defining PAR region}

\item{nir.wl}{Vector of wavelengths defining NIR region}

\item{datetime}{POSIXlt object defining the date and at which the run will 
take place. N datetime POSIXlt object defining the date and at which the run 
will take place. Note that runs at night and during the winter can give poor 
results.}

\item{history.prefix}{Prefix in histroy file name. Will be appended to 
history path.}

\item{edr.exe.name}{Name of EDR executable. Default = 'ed_2.1-opt'}

\item{change.history.time}{Logical. If `TRUE`, rename the history file to 
the time specified in `datetime`. Default = TRUE.}

\item{output.path}{Path to store all output files. Default is current 
working directory.}

\item{clean}{Logical. If `TRUE`, remove all files generated by this function 
(e.g. cloned history file, ED2IN, output HDF files).}
}
\description{
This function provides a convenient way to call the ED 
radiative transfer module (EDR, which simulates full spectral return of an 
ED patch for a given point in time) directly from R.
}

