% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.auto.R
\name{invert.auto}
\alias{invert.auto}
\title{Inversion with automatic convergence checking}
\usage{
invert.auto(observed, invert.options, return.samples = TRUE,
  save.samples = NULL, quiet = FALSE, parallel = TRUE,
  parallel.cores = NULL, parallel.output = "/dev/null")
}
\arguments{
\item{observed}{Vector, matrix, or data frame (coerced to matrix) of 
observed values. For spectral data, wavelengths are rows and spectra are 
columns. Dimensions must align with the output of `model`.}

\item{invert.options}{Parameters related to inversion.
Parameters specific to `invert.auto` are described here.
For the remaining parameters, see \code{\link{invert.custom}}.

model The model to be inverted. This should be an R function that takes 
`params` as input and returns one column of `observed` (nrows should be the 
same). Constants should be implicitly included here.

nchains Number of independent chains.

inits.function Function for randomly generating initial conditions.

ngibbs.max Maximum number of total iterations (per chain). DEFAULT = 5e6

ngibbs.min Minimum number of total iterations (per chain). DEFAULT = 5000.

ngibbs.step Number of iterations between convergence checks. Default = 1000.}

\item{return.samples}{Include full samples list in output. Default = TRUE.}

\item{save.samples}{Save samples to file as the inversion proceeds (useful for debugging).
If NULL, do not save samples. Default = NULL.}

\item{quiet}{Suppress progress bar and status messages. Default=FALSE}

\item{parallel}{Logical. Whether or not to run multiple chains in parallel on multiple cores (default = TRUE).}

\item{parallel.cores}{Number of cores to use for parallelization. If NULL (default), allocate one fewer than detected number of cores.}

\item{parallel.output}{Filename (or '' for stdout) for printing parallel outputs. Use with caution. Default = '/dev/null'.}
}
\value{
List of "results" (summary statistics) and "samples" (mcmc.list object, or "NA" if return.samples=FALSE)
}
\details{
Performs an inversion via the `invert.custom` function with multiple chains and automatic convergence checking. 
Convergence checks are performed using the multivariate Gelman-Rubin diagnostic.
}

