% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.auto.R
\name{invert.auto}
\alias{invert.auto}
\title{Inversion with automatic convergence checking}
\usage{
invert.auto(observed, invert.options, return.samples = TRUE,
  save.samples = NULL, quiet = FALSE, parallel = TRUE,
  parallel.cores = NULL, parallel.output = "/dev/null")
}
\arguments{
\item{observed}{Vector, matrix, or data frame (coerced to matrix) of
observed values. For spectral data, wavelengths are rows and spectra are
columns. Dimensions must align with the output of \code{model}.}

\item{invert.options}{Parameters related to inversion.}

\item{return.samples}{Include full samples list in output. Default = \code{TRUE.}}

\item{save.samples}{Save samples to file as the inversion proceeds (useful
for debugging). If \code{NULL}, do not save samples. Default = \code{NULL}.}

\item{quiet}{Suppress progress bar and status messages. Default=FALSE}

\item{parallel}{Logical. Whether or not to run multiple chains in parallel
on multiple cores (default = \code{TRUE}).}

\item{parallel.cores}{Number of cores to use for parallelization. If
\code{NULL} (default), allocate one fewer than detected number of cores.}

\item{parallel.output}{Filename (or '' for stdout) for printing parallel
outputs. Use with caution. Default = \code{'/dev/null'}.}
}
\value{
List including \code{results} (summary statistics), \code{samples}
(\code{mcmc.list} object, or \code{NA} if \code{return.samples=FALSE}), and other
information.
}
\description{
Inversion with automatic convergence checking
}
\details{
Performs an inversion via the \code{invert.custom} function with
multiple chains and automatic convergence checking.  Convergence checks are
performed using the multivariate Gelman-Rubin diagnostic.

Parameters specific to \code{invert.auto} are described here.
For the remaining parameters, see \code{\link[=invert.custom]{invert.custom()}}.
\itemize{
\item \code{model} -- The model to be inverted. This should be an R function that
takes \code{params} as input and returns one column of \code{observed}
(nrows should be the same). Constants should be implicitly included here.
\item \code{nchains} -- Number of independent chains.
\item \code{inits.function} -- Function for generating initial conditions.
\item \code{ngibbs.max} -- Maximum number of total iterations (per chain). DEFAULT = 5e6
\item \code{ngibbs.min} -- Minimum number of total iterations (per chain). DEFAULT = 5000.
\item \code{ngibbs.step} -- Number of iterations between convergence checks. Default = 1000.
\item \code{run_first} -- Function to run before running sampling. Takes parallel
inputs list containing runID, initial values, and resume (NULL) as an
argument.
\item \code{calculate.burnin} -- If \code{TRUE}, use \code{PEcAn.assim.batch::autoburin}
function to calculate burnin. Otherwise, assume burnin is \code{min(niter/2, iter_conv_check)}.
\item \code{threshold} -- Maximum value of the Gelman-Rubin diagnostic for
determining convergence. Default = 1.1
}
}
