% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invert.auto.R
\name{invert.auto}
\alias{invert.auto}
\title{Inversion with automatic convergence checking}
\usage{
invert.auto(observed, model, ngibbs, nchains, prior.function, inits.function,
  param.mins, burnin = 0.8 * ngibbs, n.tries = 5, return.samples = TRUE,
  target = 0.234, target.adj = 0.8, do.lsq.first = FALSE,
  do.lsq.after = 3, save.samples = NULL, ...)
}
\arguments{
\item{observed}{Matrix of observed values. Must line up with output of
'model'.}

\item{model}{Function that calls the model to be inverted. Must take a
single argument containing a vector of inputs.}

\item{ngibbs}{Number of iterations.}

\item{nchains}{Number of independent chains.}

\item{prior.function}{Function called to calculate the prior.}

\item{inits.function}{Function for randomly generating initial conditions.}

\item{param.mins}{Parameter strict lower bounds.}

\item{burnin}{Number of samples to burn-in before computing Gelman
Diagnostic. Default = 0.8 * ngibbs.}

\item{n.tries}{Number of attempted runs before giving up. Default = 5}

\item{return.samples}{Include full samples list in output. Default = TRUE.}

\item{target.adj}{Amount by which to adjust target acceptance rate every
attempt. Default = 0.8}

\item{do.lsq.first}{Initialize using least-squares optimization on first
try. Default = FALSE.}

\item{do.lsq.after}{Number of tries before starting initialization using
least-squares optimization. Default = TRUE.}

\item{save.samples}{Filename for saving samples after each iteration. If
'NULL', do not save samples. Default = NULL.}
}
\value{
List of "results" (summary statistics and Gelman Diagnostic) and
"samples"(mcmc.list object, or "NA" if return.samples=FALSE)
}
\description{
Inversion with automatic convergence checking
}
\details{
Performs an inversion via the `invert.custom` function with
multiple chains and automatic convergence checking. Convergence checks are
performed using the multivariate Gelman-Rubin diagnostic.
}

