% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.auto.R
\name{invert.auto}
\alias{invert.auto}
\title{Inversion with automatic convergence checking}
\usage{
invert.auto(observed, settings, return.samples = TRUE, save.samples = NULL,
  quiet = FALSE)
}
\arguments{
\item{observed}{Matrix of observed values. Must line up with output of 
'model'.}

\item{settings}{R list object containing the following elements:

inits Vector of initial values of model parameters to be inverted.

ngibbs Number of MCMC iterations

prior.function Function for use as prior. Should take a vector of parameters 
as input and return a single value -- the sum of their log-densities -- as 
output.

param.mins Vector of minimum values for inversion parameters

model The model to be inverted. This should be an R function that takes 
`params` as input and returns one column of `observed` (nrows should be the 
same). Constants should be implicitly included here.

adapt Number of steps for adapting covariance matrix (i.e. adapt every 'n' 
steps). Default=100
adj_min Minimum threshold for rescaling Jump standard deviation.  Default = 
0.1.

target Target acceptance rate. Default=0.234

do.lsq Perform least squares optimization first (see `invert.lsq`), and use 
outputs to initialize Metropolis Hastings. This may improve mixing time, but 
risks getting caught in a local minimum.  Default=FALSE

nchains Number of independent chains.

inits.function Function for randomly generating initial conditions.
burnin Number of samples to burn-in before computing Gelman 
Diagnostic. Default = 0.8 * ngibbs.

n.tries Number of attempted runs before giving up. Default = 5

do.lsq.first Initialize using least-squares optimization on first 
try. Default = FALSE.

do.lsq.after Number of tries before starting initialization using 
least-squares optimization. Default = TRUE.}

\item{return.samples}{Include full samples list in output. Default = TRUE.
target.adj Amount by which to adjust target acceptance rate every 
attempt. Default = 0.8}

\item{save.samples}{Filename for saving samples after each iteration. If 
'NULL', do not save samples. Default = NULL.}
}
\value{
List of "results" (summary statistics and Gelman Diagnostic) and 
"samples"(mcmc.list object, or "NA" if return.samples=FALSE)
}
\details{
Performs an inversion via the `invert.custom` function with 
multiple chains and automatic convergence checking. Convergence checks are 
performed using the multivariate Gelman-Rubin diagnostic.
}

