% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invert.lsq.R
\name{invert.lsq}
\alias{invert.lsq}
\title{Least squares model inversion}
\usage{
invert.lsq(observed, inits, model, lower = NULL, upper = NULL)
}
\arguments{
\item{observed}{Vector of observations (e.g. a reflectance spectrum).}

\item{inits}{Vector of initial conditions for the parameters.}

\item{model}{An R function that calls the RTM and returns the error to be
minimized. Be sure to include constants here.}

\item{lower}{Lower bounds on parameters (default=NULL, which means -Inf).}

\item{uppper}{Upper bounds on parameters (default=NULL, which means +Inf).}
}
\description{
Least squares model inversion
}
\details{
Performs a least-squares inversion of an arbitrary radiative
transfer model (passed as an R function). The inversion attempts to minimize
the sum of residual least squares between modeled and observed spectra via
the Levenberg-Marquardt algorithm (`nls.lm` function from the `minpack.lm`
package).
}
\author{
Alexey Shiklomanov
}

