% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invert.slow.R
\name{invert.slow}
\alias{invert.slow}
\title{Bayesian inversion of a model}
\usage{
invert.slow(observed, inits, constants, ngibbs, prior, pm, model)
}
\arguments{
\item{observed}{Vector, matrix, or data frame (coerced to matrix) of
observed values. For spectral data, wavelengths are rows and
spectra are columns.}

\item{inits}{Vector of initial values of model parameters to be
inverted.}

\item{constants}{Vector of model constants.}

\item{ngibbs}{Number of MCMC iterations}

\item{prior}{List of functions used as priors (e.g. dnorm(x, 3, 0.5, 1)).
NOTE: These should return the LOG density of the value
(i.e. log=TRUE should be set)}

\item{pm}{Vector of minimum values for inversion parameters}

\item{model}{The model to be inverted. This should be an R function
that takes `inits` and `constants` as input and returns
one column of `observed` (nrows should be the same).}
}
\description{
Bayesian inversion of a model
}
\details{
Performs a Bayesian inversion of an arbitrary function
     using the Metropolis-Hastings algorithm.
     This is substantially (~80x) slower than
     `invert.fast` because it is implemented as a basic R loop,
     but it is also more versatile and is intended for use with
     custom models.
}

