% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_fieldspec.R
\name{invert_fieldspec}
\alias{invert_fieldspec}
\title{Quickly invert a field spectrum}
\usage{
invert_fieldspec(spectra, type = "R", method = "fmm",
  prospect_version = 5, prior = prospect_bt_prior(prospect_version), ...)
}
\arguments{
\item{spectra}{Observed spectra to invert. Must be class \link{spectra}.}

\item{type}{Type of observation, one of "R", "T", "PA", or "CRR" (see Details)}

\item{method}{One of the methods for \code{\link[stats:splinefun]{stats::splinefun()}} (for polynomial
and periodic splines) or \code{\link[stats:approxfun]{stats::approxfun()}} (for constant or linear).
Default is \code{"fmm"} (same as splinefun).}

\item{prospect_version}{Version of PROSPECT to use. Default is "5B".}

\item{prior}{BayesianTools prior object.}

\item{...}{Additional arguments to \link{invert_bt}}
}
\description{
Automatically defines an RTM model based on the observed wavelengths and
spectra type.
}
\details{
Spectra type must be one of the following:
\itemize{
\item "R" -- Reflectance (default)
\item "T" -- Transmittance
\item "RT" -- Reflectance and transmittance, in separate columns. This requires
that the observation.
\item "PA" -- Pseudo-absorbance, defined as \code{log10(1 / reflectance)}
\item "CRR" -- Continuum-removed reflectance. See
\link[prospectr:continuumRemoval]{prospectr::continuumRemoval}. Requires the \code{prospectr} package.
}
}
