% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_fieldspec.R
\name{invert_fieldspec}
\alias{invert_fieldspec}
\title{Quickly invert a field spectrum}
\usage{
invert_fieldspec(spectra, type = spectra_types(spectra), method = "fmm",
  prospect_version = 5, prior = prospect_bt_prior(prospect_version),
  test = TRUE, ...)
}
\arguments{
\item{spectra}{Observed spectra to invert. Must be class \link{spectra}.}

\item{method}{One of the methods for \code{\link[stats:splinefun]{stats::splinefun()}} (for polynomial
and periodic splines) or \code{\link[stats:approxfun]{stats::approxfun()}} (for constant or linear).
Default is \code{"fmm"} (same as splinefun).}

\item{prospect_version}{Version of PROSPECT to use. Default is "5B".}

\item{prior}{BayesianTools prior object.}

\item{test}{Logical. If \code{TRUE}, test that model and observations match
before starting sampling.}

\item{...}{Additional arguments to \link{invert_bt}}
}
\description{
Automatically defines an RTM model based on the observed wavelengths and
spectra type.
}
