% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process-output.R
\name{load.from.name}
\alias{load.from.name}
\title{Load object from an RData file}
\usage{
load.from.name(filename, filepath = ".")
}
\arguments{
\item{filename}{Full name (without path!) of RData file}

\item{filepath}{Path of RData file (default=".")}
}
\description{
Load object from an RData file
}

