% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matplot.spectra.R
\name{matplot.spectra}
\alias{matplot.spectra}
\title{Plot multiple spectra on same graph}
\usage{
\method{matplot}{spectra}(spectra, ...)
}
\arguments{
\item{spectra}{Vector (\code{length = length(wavelengths)}) or matrix (\code{ncol = length(wavelengths)})}

\item{...}{Additional arguments to \code{matplot}}
}
\description{
Plot multiple spectra on same graph
}
