% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spectra.R
\name{plot.spectra}
\alias{plot.spectra}
\title{Plot spectra vs. wavelength}
\usage{
\method{plot}{spectra}(spectra, type = "l", ...)
}
\arguments{
\item{spectra}{Vector (\code{length = length(wavelengths)}) or matrix (\code{ncol = length(wavelengths)})}

\item{type}{plot style, e.g. "l" for lines, "b" for lines and points}

\item{...}{Additional arguments to \code{plot}}
}
\description{
Plot spectra vs. wavelength
}
