% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostream.R
\name{pro2s}
\alias{pro2s}
\title{Coupled PROSPECT-Two-stream model}
\usage{
pro2s(param, prospect.version = 5)
}
\arguments{
\item{param}{Model parameters, in the following order: N, Cab, (Car, Cbrown), Cw, Cm, solar zenith angle, LAI, soil_moisture}

\item{prospect.version}{Version of PROSPECT to use (4, 5, or '5B'; default=5)}
}
\description{
Coupled PROSPECT-Two-stream model
}
