% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prospect.R
\name{prospect}
\alias{prospect}
\title{PROSPECT (4, 5, or 5B) model}
\usage{
prospect(param, version = "5B")
}
\arguments{
\item{param}{Vector of PROSPECT parameter values:
N: Effective number of leaf layers (>1)
Cab: Leaf chlorophyll content (ug/cm2) (>0)
(5) Car: Leaf carotenoid content (ug/cm2) (>0)
(5B) Cbrown: Leaf brown matter content (ug/cm2) (>0)
Cw: Leaf water content (cm) (>0)
Cm: Leaf dry matter content (ug/cm2) (>0)}

\item{version}{PROSPECT version: 4, 5, or "5B" (default)}
}
\value{
Matrix (2101 x 2) of simulated reflectance (column 1)
     and transmittance (column 2) values from 400:2100 nm
}
\description{
PROSPECT (4, 5, or 5B) model
}
\details{
R wrapper for PROSPECT models
}
\author{
Alexey Shiklomanov
}

