% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prospect4}
\alias{prospect4}
\title{PROSPECT4 model}
\usage{
prospect4(param, reflectance = TRUE)
}
\arguments{
\item{param}{Vector of PROSPECT4 parameter values:
N: Effective number of leaf layers (>1)
Cab Leaf chlorophyll content (>0)
Cw Leaf water content (>0)
Cm Leaf dry matter content (>0)}

\item{reflectance}{Return reflectance (TRUE, default) or transmittance (FALSE)}
}
\value{
Vector (length 2100) of simulated reflectance/transmittance values from 400:2100 nm
}
\description{
PROSPECT4 model
}
\details{
Concise R wrapper for PROSPECT4 model
}
\author{
Alexey Shiklomanov
}

