% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prospect4}
\alias{prospect4}
\title{PROSPECT4 model}
\usage{
prospect4(N, Cab, Cw, Cm, transmittance = FALSE)
}
\arguments{
\item{N}{Effective number of leaf layers (>1)}

\item{Cab}{Leaf chlorophyll content (>0)}

\item{Cw}{Leaf water content (>0)}

\item{Cm}{Leaf dry matter content (>0)}

\item{transmittance}{Return transmittance instead of reflectance (default = FALSE)}
}
\value{
Vector (length 2100) of simulated reflectance/transmittance values from 400:2100 nm
}
\description{
PROSPECT4 model
}
\author{
Alexey Shiklomanov
}

