% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestools.R
\name{prospect_bt_prior}
\alias{prospect_bt_prior}
\title{Quick BayesianTools prior creator for PROSPECT model}
\usage{
prospect_bt_prior(version, custom_prior = list())
}
\description{
Quick BayesianTools prior creator for PROSPECT model
}
