% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rsr.R
\name{read.rsr.folder}
\alias{read.rsr.folder}
\title{Read and process RSR data from directory}
\usage{
read.rsr.folder(dir.path, type)
}
\arguments{
\item{dir.path}{Directory containing RSR data}

\item{type}{Type of sensor. Options are: landsat, avhrr}
}

