% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{resample.default}
\alias{resample.matrix}
\alias{resample.spectra}
\title{Resample vector, matrix, or spectra}
\usage{
resample(...)

\method{resample}{default}(values, from, to, ...)

\method{resample}{matrix}(values, from, to, ...)

\method{resample}{spectra}(spectra, to, ...)
}
\arguments{
\item{...}{Additional arguments to \code{\link[stats:splinefun]{stats::splinefun()}}}

\item{values}{Vector or matrix of values. Length of vector, or \code{nrow} of
matrix must match length of \code{from}.}

\item{from}{X values for interpolation (for \code{spectra} objects, this is
assumed to be the \code{wavelengths} attribute.)}

\item{to}{Y values onto which to interpolate. For \code{spectra} objects, this
should be new wavelengths.}

\item{spectra}{An object of class \code{spectra}. See \code{\link[=spectra]{spectra()}} for more
details.}
}
\value{
Object of the same class as \code{values}, resampled to the \code{to} values.
}
\description{
Conveinent wrapper around base R's \code{splinefun}. See \code{\link[stats:splinefun]{stats::splinefun()}}
documentation for information on different spline methods and additional
arguments.
}
