% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestools.R
\name{rtm_loglike}
\alias{rtm_loglike}
\title{Generic log-likelihood generator for RTMs}
\usage{
rtm_loglike(nparams, model, observed, lag.max = NULL, verbose = TRUE,
  ...)
}
\description{
Generic log-likelihood generator for RTMs
}
