% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectral-response.R
\docType{data}
\name{sensor.list}
\alias{sensor.list}
\title{Sensor spectral response functions}
\format{\preformatted{ chr [1:11] "identity" "aviris.ng" "aviris.classic" "hyperion" "chris.proba" "landsat5" "landsat7" "landsat8" "modis" "viirs" "avhrr"
}}
\usage{
sensor.list
}
\description{
Sensor spectral response functions
}
\keyword{datasets}

