% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectra}
\alias{spectra}
\alias{is_spectra}
\title{Spectra S3 class}
\usage{
spectra(spectra, wavelengths = 400:2500)

is_spectra(spectra)
}
\arguments{
\item{spectra}{Vector (\code{length = length(wavelengths)}) or matrix (\code{ncol = length(wavelengths)})}

\item{wavelengths}{Wavelengths of spectra.}
}
\description{
Spectra S3 class
}
\section{Functions}{
\itemize{
\item \code{is_spectra}: Test if object is class \code{spectra}
}}

