% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.spectra.R
\name{str.spectra}
\alias{str.spectra}
\title{Structure of \code{spectra} object}
\usage{
\method{str}{spectra}(spectra, ...)
}
\arguments{
\item{spectra}{Object of class \code{spectra}}

\item{...}{additional arguments, currently ignored}
}
\description{
Structure of \code{spectra} object
}
