% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.spectra.R
\name{[[<-.spectra}
\alias{[[<-.spectra}
\title{Assign values to spectra by wavelength}
\usage{
\method{[[}{spectra}(spectra, wavelength, j) <- value
}
\arguments{
\item{spectra}{Object of class \code{spectra}}

\item{wavelength}{Wavelength vector to select}

\item{j}{index specifying elements to extract or replace.
See \code{\link[base:Extract]{base::Extract}} for details.}

\item{value}{Vector or matrix of values to assign}
}
\description{
Assign values to spectra by wavelength
}
