% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process-output.R
\name{summary.simple}
\alias{summary.simple}
\title{Simple summary statistics on MCMC samples}
\usage{
\method{summary}{simple}(samples)
}
\arguments{
\item{samples}{Matrix of MCMC samples}
}
\description{
Calculate simple univariate summary statistics and return as
named list
}

